More minor C++ style modifications

+++ lte/gateway/c/session_manager/SessionState.cpp
@@ -195,7 +195,7 @@ SessionState::SessionState(
     const std::string& imsi, const std::string& session_id,
     const SessionConfig& cfg, StaticRuleStore& rule_store,
     const magma::lte::TgppContext& tgpp_context, uint64_t pdp_start_time,
-    const CreateSessionResponse& csr, const int shard_id)
+    const CreateSessionResponse& csr, const uint16_t shard_id)
     : imsi_(imsi),
       session_id_(session_id),
       // Request number set to 1, because request 0 is INIT call
