+++ lte/gateway/c/session_manager/SessionState.h
@@ -136,7 +136,7 @@ class SessionState {
       const std::string& imsi, const std::string& session_id,
       const SessionConfig& cfg, StaticRuleStore& rule_store,
       const magma::lte::TgppContext& tgpp_context, uint64_t pdp_start_time,
-      const CreateSessionResponse& csr, const int shard_id);
+      const CreateSessionResponse& csr, const uint16_t shard_id);
 
   SessionState(
       const StoredSessionState& marshaled, StaticRuleStore& rule_store);
